packages/sdk-python/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "direct-method-mqtt-python"
version = "1.0.0"
description = "MQTT client implementation for Python - part of Direct Method Library"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Direct Method Library Team"}
]
keywords = ["mqtt", "client", "python", "iot", "messaging"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
    "Topic :: System :: Networking",
    "Topic :: Communications",
]
requires-python = ">=3.9"
dependencies = [
    "paho-mqtt>=1.6.0",
]

[project.urls]
Homepage = "https://github.com/ivayloivanov7/direct-method-library"
Repository = "https://github.com/ivayloivanov7/direct-method-library.git"
Issues = "https://github.com/ivayloivanov7/direct-method-library/issues"
Documentation = "https://github.com/ivayloivanov7/direct-method-library/tree/main/packages/sdk-python"

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "types-paho-mqtt",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"direct_method_mqtt" = ["py.typed"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["direct_method_mqtt"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
show_error_codes = true
exclude = [
    "build/",
    "dist/",
]

[[tool.mypy.overrides]]
module = "paho.mqtt.*"
ignore_missing_imports = true